requirements.txt
mysql-connector-python
bcrypt
# txt